refactor: simplify ACP reply destination parsing
