+++ .agents/scripts/_team_interface_acp_reply.mjs
@@ -1,157 +1,14 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2025-2026 Marcus Quinn
 
+import {
+  DIRECT_GREETING_REPLY,
+  trustedReplyDestination,
+} from "./_team_interface_acp_destination.mjs";
+
 const MAX_REPLY_BYTES = 16 * 1024;
-const UUID_PATTERN = "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}";
-const EVENT_ID_PATTERN = "[0-9a-f]{64}";
 const REPLY_OPEN = "<buzz-reply>";
 const REPLY_CLOSE = "</buzz-reply>";
-const DIRECT_GREETING_REPLY = "Hi! What would you like to work on?";
-
-function promptBlocks(message) {
-  const prompt = message?.params?.prompt;
-  if (!Array.isArray(prompt)) {
-    return null;
-  }
-  return prompt
-    .filter((block) => block && typeof block === "object" && typeof block.text === "string")
-    .map((block) => block.text);
-}
-
-function isDirectGreeting(message) {
-  const prompt = message?.params?.prompt;
-  if (!Array.isArray(prompt)) {
-    return false;
-  }
-  const contextBlock = prompt.find(
-    (block) => typeof block?.text === "string" && block.text.startsWith("[Context]"),
-  );
-  const eventBlock = [...prompt].reverse().find(
-    (block) => typeof block?.text === "string" && block.text.startsWith("[Buzz event:"),
-  );
-  if (!contextBlock?.text.match(/^Scope: dm$/im) || !eventBlock) {
-    return false;
-  }
-  const eventLines = eventBlock.text.split(/\r?\n/u);
-  const contentIndex = eventLines.findIndex((line) => line.startsWith("Content: "));
-  const tagsIndex = eventLines.findLastIndex((line) => line.startsWith("Tags: "));
-  if (contentIndex < 0 || tagsIndex !== contentIndex + 1) {
-    return false;
-  }
-  const content = eventLines[contentIndex].slice("Content: ".length).trim();
-  return /^(?:hi|hello|hey|hiya|howdy|good (?:morning|afternoon|evening))[!.?]*$/iu.test(
-    content || "",
-  );
-}
-
-function trustedReplyInstructions(eventLines, contentIndex, tagsIndex) {
-  const instructionPattern = new RegExp(
-    "^IMPORTANT: [^\\r\\n]*use `--reply-to (" + EVENT_ID_PATTERN + ")`[^\\r\\n]*$",
-    "i",
-  );
-  const instructions = [];
-  for (const [index, line] of eventLines.entries()) {
-    const trustedEnvelope = contentIndex < 0 || index < contentIndex ||
-      (tagsIndex > contentIndex && index > tagsIndex);
-    const instruction = line.match(instructionPattern);
-    if (!instruction) {
-      if (trustedEnvelope && line.startsWith("IMPORTANT: ") && line.includes("`--reply-to ")) {
-        return null;
-      }
-      continue;
-    }
-    if (!trustedEnvelope) {
-      return null;
-    }
-    instructions.push(instruction);
-  }
-  return instructions.length <= 1 ? instructions : null;
-}
-
-function trustedReplyDestination(message) {
-  const blocks = promptBlocks(message);
-  if (!blocks || blocks.length === 0) {
-    return null;
-  }
-  const contextIndexes = blocks
-    .map((block, index) => block.startsWith("[Context]") ? index : -1)
-    .filter((index) => index >= 0);
-  if (contextIndexes.length !== 1) {
-    return null;
-  }
-  const contextIndex = contextIndexes[0];
-  const contextLines = blocks[contextIndex].split(/\r?\n/u);
-  const eventMarker = contextLines.indexOf("[Event]");
-  const contextHeader = eventMarker < 0 ? contextLines : contextLines.slice(0, eventMarker);
-  const channelLines = contextHeader.filter((line) => line.startsWith("Channel: "));
-  const channel = channelLines.length === 1
-    ? channelLines[0].match(new RegExp(`^Channel: [^\\r\\n]* \\(#(${UUID_PATTERN})\\)$`, "i"))
-    : null;
-  const eventBlocks = [];
-  if (eventMarker >= 0) {
-    eventBlocks.push({lines: contextLines.slice(eventMarker + 1), structured: false});
-  }
-  for (const [index, block] of blocks.entries()) {
-    if (block.startsWith("[Buzz event:")) {
-      if (index <= contextIndex) {
-        return null;
-      }
-      eventBlocks.push({lines: block.split(/\r?\n/u), structured: true});
-    }
-  }
-  if (!channel || eventBlocks.length !== 1) {
-    return null;
-  }
-  const {lines: eventLines, structured} = eventBlocks[0];
-  const contentIndexes = eventLines
-    .map((line, index) => line.startsWith("Content: ") ? index : -1)
-    .filter((index) => index >= 0);
-  const tagsIndexes = eventLines
-    .map((line, index) => line.startsWith("Tags: ") ? index : -1)
-    .filter((index) => index >= 0);
-  if (
-    structured &&
-    (contentIndexes.length !== 1 || tagsIndexes.length !== 1 || tagsIndexes[0] <= contentIndexes[0])
-  ) {
-    return null;
-  }
-  const contentIndex = contentIndexes[0] ?? -1;
-  const tagsIndex = tagsIndexes.at(-1) ?? -1;
-  const eventHeader = eventLines.slice(0, contentIndex < 0 ? undefined : contentIndex);
-  const eventIdLines = eventHeader.filter((line) => line.startsWith("Event ID: "));
-  const eventId = eventIdLines.length === 1
-    ? eventIdLines[0].match(new RegExp(`^Event ID: (${EVENT_ID_PATTERN})$`, "i"))
-    : null;
-  if (!eventId) {
-    return null;
-  }
-  const eventChannelLines = eventHeader.filter((line) => line.startsWith("Channel: "));
-  const eventChannel = eventChannelLines.length === 1
-    ? eventChannelLines[0].match(
-      new RegExp(`^Channel: [^\\r\\n]* \\(#(${UUID_PATTERN})\\)$`, "i"),
-    )
-    : null;
-  if (
-    eventChannelLines.length > 1 ||
-    (eventChannelLines.length === 1 && (!eventChannel || eventChannel[1] !== channel[1]))
-  ) {
-    return null;
-  }
-  const instructedReplies = trustedReplyInstructions(eventLines, contentIndex, tagsIndex);
-  if (!instructedReplies) {
-    return null;
-  }
-  const structuredFallback = structured && contentIndex >= 0 && tagsIndex > contentIndex &&
-    eventChannelLines.length === 1;
-  if (instructedReplies.length === 0 && !structuredFallback) {
-    return null;
-  }
-  return {
-    channel: channel[1],
-    directGreeting: isDirectGreeting(message),
-    replyTo: instructedReplies[0]?.[1] ?? eventId[1],
-  };
-}
 
 function sanitizeReply(text) {
   return text
